requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-core
langchain-chroma
langchain-huggingface
langchain-google-genai
huggingface-hub
sentence-transformers
chromadb
PyMuPDF
python-dotenv
protobuf
gunicorn